Move prebuild utils
